# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Commerce OS API — Phase 2
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List all warehouses
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create a new warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}:
    get:
      operationId: getWarehouse
      tags: [warehouses]
      summary: Get a warehouse by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "404":
          description: Not found
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update a warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseUpdate"
      responses:
        "200":
          description: Updated warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete a warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Inventory ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Warehouse:
      type: object
      required: [id, name, isDefault, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        isDefault:
          type: boolean
        isActive:
          type: boolean
        inventoryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    WarehouseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
    WarehouseUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean

    # ─── Inventory ─────────────────────────────────────────────────────────
...